Revert "refactor: split browser AgentCapabilities from pydantic-ai capabilities"
This reverts commit c4ced481a1baaa7c66fed561f2df54f4cb1c9bb3.
