+++ src/phoenix/server/agents/dependencies.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 
 from pydantic_ai.mcp import MCPServerStreamableHTTP
 
-from phoenix.server.agents.agent_capabilities import AgentCapabilities
+from phoenix.server.agents.capabilities import AgentCapabilities
 from phoenix.server.agents.context import ResolvedContexts
